Pull request 548: [Android] Change upstreamTimeoutMs from int to long
Merge in ADGUARD-CORE-LIBS/dns-libs from fix/androud_upstream_timeout_ms_long to master

Squashed commit of the following:

commit 2b3192b20a7ae83607021a0b5e083e04ca5ab4c7
Author: Sergey Fionov <sfionov@adguard.com>
Date:   Tue Jul 11 15:37:19 2023 +0300

    Fix instrumented tests

commit 1891cfccbe94032132682f05a9744a74c1b8494b
Author: Sergey Fionov <sfionov@adguard.com>
Date:   Tue Jul 11 15:25:57 2023 +0300

    Fix build

commit d81893b7ef78df093cf901d448a63d2448716a56
Author: Sergey Fionov <sfionov@adguard.com>
Date:   Tue Jul 11 15:17:17 2023 +0300

    Change upstreamTimeoutMs from int to long

+++ platform/android/dnsproxy/lib/src/androidTest/java/com/adguard/dnslibs/proxy/DnsProxyTest.java
@@ -475,7 +475,7 @@ public class DnsProxyTest {
 
     @Test
     public void testTestUpstream() {
-        final int timeout = 500; // ms
+        final long timeout = 500; // ms
         IllegalArgumentException e0 = null;
         try {
             DnsProxy.testUpstream(new UpstreamSettings("123.12.32.1:1493", new ArrayList<String>(), null, 42),
